+++ src/SketchSolver/InputParser/BooleanDAGCreator.h
@@ -76,7 +76,7 @@ public:
 
 
 	string new_name() {
-		int p = 0;
+		size_t p = 0;
 		char tmpbuf[200];
 		tmpbuf[p] = 't'; p++;
 		writeInt(tmpbuf, new_names, p);
